Detach SampledRelation from SampleNode
