+++ presto-main/src/main/java/com/facebook/presto/sql/planner/optimizations/SampleNodeUtil.java
@@ -1,0 +1,34 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.sql.planner.optimizations;
+
+import com.facebook.presto.sql.planner.plan.SampleNode.Type;
+import com.facebook.presto.sql.tree.SampledRelation;
+
+public final class SampleNodeUtil
+{
+    private SampleNodeUtil() {}
+
+    public static Type fromType(SampledRelation.Type sampleType)
+    {
+        switch (sampleType) {
+            case BERNOULLI:
+                return Type.BERNOULLI;
+            case SYSTEM:
+                return Type.SYSTEM;
+            default:
+                throw new UnsupportedOperationException("Unsupported sample type: " + sampleType);
+        }
+    }
+}
